requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
pillow==11.1.0
slug==2.0
slugify==0.0.1
sqlparse==0.5.3
